Revert "Tun4PacketTes"
This reverts commit 6fc46f4cda8e314d96cbc4389d964a97f7d8ffb9.
