+++ transport/src/main/java/io/netty/channel/socket/Tun6Packet.java
@@ -18,7 +18,6 @@ package io.netty.channel.socket;
 import io.netty.buffer.ByteBuf;
 import io.netty.util.internal.StringUtil;
 
-import java.net.Inet6Address;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 
@@ -77,7 +76,7 @@ public class Tun6Packet extends TunPacket {
             try {
                 byte[] dst = new byte[INET6_SOURCE_ADDRESS_LENGTH];
                 content().getBytes(INET6_SOURCE_ADDRESS, dst, 0, INET6_SOURCE_ADDRESS_LENGTH);
-                sourceAddress = Inet6Address.getByAddress(dst);
+                sourceAddress = InetAddress.getByAddress(dst);
             } catch (UnknownHostException e) {
                 // unreachable code
                 throw new IllegalStateException();
@@ -93,7 +92,7 @@ public class Tun6Packet extends TunPacket {
             try {
                 byte[] dst = new byte[INET6_SOURCE_ADDRESS_LENGTH];
                 content().getBytes(INET6_DESTINATION_ADDRESS, dst, 0, INET6_DESTINATION_ADDRESS_LENGTH);
-                destinationAddress = Inet6Address.getByAddress(dst);
+                destinationAddress = InetAddress.getByAddress(dst);
             } catch (UnknownHostException e) {
                 // unreachable code
                 throw new IllegalStateException();
